requirements.txt
alabaster==0.7.12
appdirs==1.4.4
asgiref==3.3.1
Babel==2.9.0
certifi==2020.12.5
chardet==4.0.0
distlib==0.3.1
Django==3.0.1
django-crispy-forms==1.11.2
django-etc==1.3.0
django-extensions==3.1.1
django-filter==2.4.0
django-hitcount==1.3.3
django-password-validation==0.1.1
django-timedeltafield==0.7.10
djangorestframework==3.12.4
docutils==0.16
filelock==3.0.12
idna==2.10
imagesize==1.2.0
Jinja2==2.11.3
Markdown==3.3.4
MarkupSafe==1.1.1
none==0.1.0a2
packaging==20.9
Pillow==8.2.0
psycopg2-binary==2.8.6
Pygments==2.8.1
pyparsing==2.4.7
pytz==2021.1
requests==2.25.1
rfc3986==1.4.0
six==1.15.0
snowballstemmer==2.1.0
Sphinx==3.5.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sphinxcontrib-websupport==1.2.4
sqlparse==0.4.1
urllib3==1.26.4
validate-email==1.3
virtualenv==20.4.2
